Add marks (#35)
* refactor: заменить лайки на оценки для фильмов

* feat: добавить SqlHelper

---------

+++ src/main/java/ru/yandex/practicum/filmorate/model/Review.java
@@ -31,5 +31,5 @@ public class Review {
     Long filmId;
 
     @Builder.Default
-    Long useful = 0L;
+    Integer useful = 0;
 }
